Fix state pattern
